# backend/api-specification.yaml
openapi: 3.0.3
info:
  title: Planning Poker API
  description: |
    A comprehensive REST API for collaborative agile story estimation with real-time features.
    
    ## Features
    - Multiple sizing methodologies (Fibonacci, T-Shirt, Powers of 2, Linear, Custom)
    - Real-time collaboration with WebSocket support
    - Team management and user tracking
    - Story backlog management
    - Vote confidence tracking and analytics
    - Session history and reporting
    - Export/Import capabilities
    - Timer and auto-reveal functionality
    
    ## Authentication
    All write operations and sensitive reads require a JWT Bearer token obtained from
    `createSession` or `joinSession`. Include it as:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - 100 requests per minute per IP for session creation
    - 1000 requests per minute for other endpoints
    
  version: 1.0.0
  contact:
    name: PandaC Team
    email: support@pandac.com
    url: https://github.com/pandac/planning-poker
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.planningpoker.pandac.com/api
    description: Production server
...
paths:
...
  /sessions/{sessionCode}/stories:
    get:
      tags: [Stories]
      summary: List all stories
      description: Get all stories in the session backlog
      operationId: getStories
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - name: status
          in: query
          description: Filter by story status
          schema:
            $ref: '#/components/schemas/StoryStatus'
      responses:
        '200':
          description: List of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Story'
    
    post:
      tags: [Planning Poker]
      summary: Create a new story
      description: Add a story to the backlog
      operationId: createStory
      parameters:
        - $ref: '#/components/parameters/SessionCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoryRequest'
      responses:
        '201':
          description: Story created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
  /sessions/{sessionCode}/stories/{storyId}:
    get:
      tags: [Stories]
      summary: Get story details
      operationId: getStory
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
      responses:
        '200':
          description: Story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
    
    put:
      tags: [Stories]
      summary: Update story
      operationId: updateStory
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoryRequest'
      responses:
        '200':
          description: Story updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
    
    delete:
      tags: [Stories]
      summary: Delete story
      operationId: deleteStory
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
      responses:
        '204':
          description: Story deleted
  /sessions/{sessionCode}/stories/{storyId}/finalize:
    post:
      tags: [Stories]
      summary: Finalize story estimate
      description: Set the final agreed estimate for a story
      operationId: finalizeEstimate
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [finalEstimate]
              properties:
                finalEstimate:
                  type: string
                  description: The final agreed estimate
                  example: "5"
                notes:
                  type: string
                  description: Optional notes about the estimate
      responses:
        '200':
          description: Estimate finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
  /sessions/{sessionCode}/stories/{storyId}/reset:
    post:
      tags: [Stories]
      summary: Reset story for revoting
      description: Clear estimate and all votes to allow revoting (moderator only)
      operationId: resetStory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
      responses:
        '200':
          description: Story reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== VOTING ENDPOINTS ====================
  /sessions/{sessionCode}/stories/{storyId}/votes:
    get:
      tags: [Votes]
      summary: Get votes for a story
      operationId: getVotes
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
        - name: revealed
          in: query
          description: Only show if votes are revealed
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of votes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vote'
    
    post:
      tags: [Votes]
      summary: Cast or update vote
      description: Submit an estimate for a story
      operationId: castVote
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '201':
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
  /sessions/{sessionCode}/stories/{storyId}/votes/{userId}:
    delete:
      tags: [Votes]
      summary: Remove vote
      description: Delete a user's vote for a story
      operationId: deleteVote
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - $ref: '#/components/parameters/StoryId'
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Vote deleted

  # ==================== USER ENDPOINTS ====================
  /sessions/{sessionCode}/users:
    get:
      tags: [Users]
      summary: Get active users
      description: List all participants in the session
      operationId: getUsers
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - name: activeOnly
          in: query
          description: Filter to only active users
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /sessions/{sessionCode}/users/{userId}:
    get:
      tags: [Users]
      summary: Get user details
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    
    put:
      tags: [Users]
      summary: Update user profile
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionCode'
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== ANALYTICS ENDPOINTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `createSession` (moderator) or `joinSession` (participant).
        Include as: `Authorization: Bearer <token>`
  parameters:
    SessionCode:
      name: sessionCode
      in: path
      required: true
      description: Unique 6-character session code
      schema:
        type: string
        pattern: '^[A-Z0-9]{6}$'
        example: "ABC123"
    StoryId:
      name: storyId
      in: path
      required: true
      description: Story identifier
      schema:
        type: integer
        format: int64
        example: 42
  schemas:
...
    CreateStoryRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "User login functionality"
        description:
          type: string
          maxLength: 2000
          example: "As a user, I want to log in securely..."
        acceptanceCriteria:
          type: string
          maxLength: 2000
        tags:
          type: array
          items:
            type: string
          example: ["frontend", "security"]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          default: MEDIUM
    UpdateStoryRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        acceptanceCriteria:
          type: string
          maxLength: 2000
        tags:
          type: array
          items:
            type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
    Story:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        acceptanceCriteria:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        status:
          $ref: '#/components/schemas/StoryStatus'
        orderIndex:
          type: integer
          description: Position in backlog
        finalEstimate:
          type: string
          description: Agreed estimate value
        estimateNotes:
          type: string
        voteCount:
          type: integer
          description: Number of votes cast
        consensus:
          type: boolean
          description: Whether all votes agree
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ========== VOTING SCHEMAS ==========
    VoteRequest:
      type: object
      required:
        - userId
        - estimate
      properties:
        userId:
          type: integer
          format: int64
          description: ID of the user casting the vote
        estimate:
          type: string
          description: Vote value (depends on sizing method)
          example: "5"
        confidence:
          type: integer
          minimum: 1
          maximum: 5
          description: Confidence level (1-5)
          example: 4
    Vote:
      type: object
      properties:
        id:
          type: integer
          format: int64
        storyId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userAvatar:
          type: string
        estimate:
          type: string
        confidence:
          type: integer
          minimum: 1
          maximum: 5
        votedAt:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        avatar:
          type: string
        isObserver:
          type: boolean
        isModerator:
          type: boolean
        isActive:
          type: boolean
        hasVoted:
          type: boolean
          description: Has voted for current story
        joinedAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        avatar:
          type: string

    # ========== ANALYTICS SCHEMAS ==========
...
    StoryStatus:
      type: string
      enum:
        - NOT_ESTIMATED
        - IN_PROGRESS
        - COMPLETED
        - SKIPPED
...
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: "2026-02-03T20:30:00Z"
            status: 404
            error: "Not Found"
            message: "Session not found"
            path: "/api/sessions/XYZ789"
    Forbidden:
      description: Access forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: "2026-02-03T20:30:00Z"
            status: 403
            error: "Forbidden"
            message: "Only moderator can perform this action"
            path: "/api/sessions/ABC123"
...